+++ core-ng-test/src/main/java/core/framework/test/module/MockFactoryImpl.java
@@ -5,10 +5,14 @@ import core.framework.api.queue.MessagePublisher;
 import core.framework.api.redis.Redis;
 import core.framework.impl.module.MockFactory;
 import core.framework.impl.queue.MessageValidator;
+import core.framework.impl.search.ElasticSearch;
 import core.framework.test.mongo.MockMongo;
 import core.framework.test.queue.MockMessagePublisher;
 import core.framework.test.redis.MockRedis;
+import core.framework.test.search.MockElasticSearch;
 import org.mockito.Mockito;
+
+import java.nio.file.Path;
 
 /**
  * @author neo
@@ -21,6 +25,7 @@ public final class MockFactoryImpl implements MockFactory {
         if (MessagePublisher.class.equals(instanceClass))
             return (T) new MockMessagePublisher<>((String) params[0], (MessageValidator) params[1]);
         if (Mongo.class.equals(instanceClass)) return (T) new MockMongo();
+        if (ElasticSearch.class.equals(instanceClass)) return (T) new MockElasticSearch((Path) params[0]);
         return Mockito.mock(instanceClass);
     }
 }