+++ core-ng/src/main/java/core/framework/impl/search/ElasticSearch.java
@@ -12,36 +12,27 @@ import org.elasticsearch.common.settings.Settings;
 import org.elasticsearch.common.transport.InetSocketTransportAddress;
 import org.elasticsearch.common.transport.TransportAddress;
 import org.elasticsearch.common.unit.TimeValue;
-import org.elasticsearch.node.Node;
-import org.elasticsearch.node.NodeBuilder;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.net.InetSocketAddress;
-import java.nio.file.Path;
 import java.time.Duration;
 import java.util.List;
 
 /**
  * @author neo
  */
-public final class ElasticSearch {
+public class ElasticSearch {
     private final Logger logger = LoggerFactory.getLogger(ElasticSearch.class);
-    private final List<TransportAddress> remoteAddresses = Lists.newArrayList();
-    private Path localHomePath;
+    private final List<TransportAddress> addresses = Lists.newArrayList();
     private Duration timeout = Duration.ofSeconds(10);
     private Duration slowOperationThreshold = Duration.ofSeconds(5);
 
     private Client client;
 
-    public void remote(String host) {
-        if (client != null) throw new Error("remote() must be called before creating client");
-        remoteAddresses.add(new InetSocketTransportAddress(new InetSocketAddress(host, 9300)));
-    }
-
-    public void local(Path localHomePath) {
-        if (client != null) throw new Error("local() must be called before creating client");
-        this.localHomePath = localHomePath;
+    public void host(String host) {
+        if (client != null) throw new Error("host() must be called before creating client");
+        addresses.add(new InetSocketTransportAddress(new InetSocketAddress(host, 9300)));
     }
 
     public void slowOperationThreshold(Duration slowOperationThreshold) {
@@ -116,32 +107,13 @@ public final class ElasticSearch {
 
     public Client client() {
         if (client == null) {
-            if (!remoteAddresses.isEmpty()) {
-                client = createRemoteElasticSearch();
-            } else if (localHomePath != null) {
-                client = createLocalElasticSearch();
-            } else {
-                throw new Error("failed to create elastic search client, please specify either remote or local");
-            }
+            client = createClient();
         }
         return client;
     }
 
-    private Client createLocalElasticSearch() {
-        StopWatch watch = new StopWatch();
-        try {
-            Settings.Builder settings = Settings.settingsBuilder()
-                .put("http.enabled", "false")       // refer to org.elasticsearch.node.Node.start()
-                .put("script.inline", "on")         // refer to https://www.elastic.co/guide/en/elasticsearch/reference/current/_setting_changes.html#migration-script-settings
-                .put("path.home", localHomePath);   // refer to org.elasticsearch.env.Environment.Environment()
-            Node node = NodeBuilder.nodeBuilder().settings(settings).local(true).node();
-            return node.client();
-        } finally {
-            logger.info("create local elastic search client, dataPath={}, elapsedTime={}", localHomePath, watch.elapsedTime());
-        }
-    }
-
-    private Client createRemoteElasticSearch() {
+    protected Client createClient() {
+        if (addresses.isEmpty()) throw new Error("addresses must not be empty, please check config");
         StopWatch watch = new StopWatch();
         try {
             Settings.Builder settings = Settings.settingsBuilder()
@@ -149,10 +121,10 @@ public final class ElasticSearch {
                 .put("client.transport.ping_timeout", new TimeValue(timeout.toMillis()))
                 .put("client.transport.ignore_cluster_name", "true");     // refer to https://www.elastic.co/guide/en/elasticsearch/client/java-api/current/transport-client.html
             TransportClient client = TransportClient.builder().settings(settings).build();
-            remoteAddresses.forEach(client::addTransportAddress);
+            addresses.forEach(client::addTransportAddress);
             return client;
         } finally {
-            logger.info("create remote elastic search client, addresses={}, elapsedTime={}", remoteAddresses, watch.elapsedTime());
+            logger.info("create elasticsearch client, addresses={}, elapsedTime={}", addresses, watch.elapsedTime());
         }
     }
 }